# 处理.dev环境

+++ eagle-core/src/main/resources/db/migration/V20200211.010__V0.1_table__wx.sql
@@ -206,7 +206,7 @@ CREATE TABLE `wx_mp_user_tags`
     `id`     BIGINT AUTO_INCREMENT,
     `appid`  VARCHAR(32) NOT NULL
         COMMENT '开发者ID(AppID)',
-    `tag_id` LONG        NOT NULL
+    `tag_id` INT(10)     NOT NULL
         COMMENT '标签ID(来着微信)',
     `name`   VARCHAR(32) NOT NULL
         COMMENT '标签名称(来着微信)',
@@ -219,7 +219,7 @@ DROP TABLE IF EXISTS `wx_mp_user_tags_relation`;
 CREATE TABLE `wx_mp_user_tags_relation`
 (
     `id`      BIGINT AUTO_INCREMENT,
-    `tags_id` LONG        NOT NULL
+    `tags_id` BIGINT      NOT NULL
         COMMENT '用户标签ID(来着自己)',
     `openid`  VARCHAR(32) NOT NULL
         COMMENT '用户的标识，对当前公众号唯一',
